chore: rename upload to upload page
